add editUser, state, constants... Upgrat details,creatModals

+++ src/module/createPage/createComponent/userComponent/formPreview.js
@@ -6,31 +6,25 @@ import deletePhoto from "../../../api/apiUser/photoApi/deletePhoto.js";
 function formPreview(id, e) {
 
   const formPreviewDiv = document.getElementById('formPreview');
-  formPreviewDiv.innerHTML = `
+  const idPreview = formPreviewDiv.dataset.id;
+
+  if (+id === +idPreview) {
+    formPreviewDiv.innerHTML = `
     <img src="${e.target.result}" alt="Photo">
     <button type="button" class="btn-close delete-photo" aria-label="Close"></button>
     `;
+  }
 
   const deletePhotoBtn = document.querySelector('.delete-photo');
-  const filePreviewImg = document.querySelector('.file__preview img');
 
-  filePreviewImg.addEventListener('mouseenter', e => {
+  if (deletePhotoBtn) {
+    deletePhotoBtn.addEventListener('click', event => {
+      event.preventDefault();
 
-    deletePhotoBtn.style.opacity = '0';
+      deletePhoto(id, formPreviewDiv);
+    });
+  }
 
-  });
-  filePreviewImg.addEventListener('mouseout', e => {
-
-    deletePhotoBtn.style.opacity = '.5';
-
-  });
-
-
-  deletePhotoBtn.addEventListener('click', event => {
-    event.preventDefault();
-
-    deletePhoto(id, formPreviewDiv);
-  });
